% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_ex}
\alias{catch_ex}
\title{Example catch data}
\format{
A data frame with 56 rows and five variables.
\describe{
\item{region}{abbreviated catch region}
\item{month_n}{month (numeric)}
\item{year}{year}
\item{catch}{catch in pieces}
\item{offset}{log(effort) used as an offset when fitting model (i.e.
coefficient fixed at one); effort units are vessel days}
}
}
\source{
Fisheries and Oceans Canada South Coast Area Office
}
\usage{
catch_ex
}
\description{
Subset of catch and effort data from west-coast Vancouver Island commercial
troll fishery for Chinook salmon.
}
\keyword{datasets}
