% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp_ex}
\alias{comp_ex}
\title{Example composition data}
\format{
A data frame with 1329 rows and seven variables.
\describe{
\item{sample_id}{sampling event id; here unique to a year day-catch region-
year}
\item{region}{abbreviated catch region}
\item{year}{year}
\item{month_n}{month (numeric)}
\item{agg}{stock aggregate; here corresponds to regional aggregates
identified by Pacific Salmon Treaty (e.g., PSD = Puget Sound)}
\item{nn}{number of samples (individual fish) collected during a sampling
event}
\item{prob}{summed individual stock assignment probabilities associated
with a sampling event}
}
}
\source{
Fisheries and Oceans Canada South Coast Area Office
}
\usage{
comp_ex
}
\description{
Subset of genetic stock composition data from west-coast Vancouver Island
commercial troll fishery for Chinook salmon.
}
\keyword{datasets}
