% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_DEFUNCT.R
\name{fit_stockseason}
\alias{fit_stockseason}
\title{Fit model}
\usage{
fit_stockseason(
  comp_dat,
  catch_dat = NULL,
  model_type = c("composition", "integrated"),
  random_walk = TRUE,
  silent = TRUE,
  optim_fix_inits = TRUE,
  nlminb_loops = 1
)
}
\arguments{
\item{comp_dat}{Composition data}

\item{catch_dat}{Catch data}

\item{model_type}{Character specifying whether inputs for a composition-only
or integrated model are being generated.}

\item{random_walk}{Logical (defaults to TRUE) specifying whether random
intercepts for year follow a random walk or are independent and identically
distributed.}

\item{silent}{Logical}

\item{optim_fix_inits}{Logical (defaults to TRUE) specifying that a
fixed-effects only version of the model is fit first to generate initial
parameters.}

\item{nlminb_loops}{How many times to run \code{\link[stats:nlminb]{stats::nlminb()}} optimization.
Sometimes restarting the optimizer at the previous best values aids
convergence. If the maximum gradient is still too large,
try increasing this to \code{2}.}
}
\value{
Output from \code{\link[TMB:sdreport]{TMB::sdreport()}}
}
\description{
Fit model
}
\examples{
m1 <- fit_stockseason(comp_dat = comp_ex, model_type = "composition")
theta <- as.list(m1, "Estimate")
theta_se <- as.list(m1, "Std. Error")
theta

m2 <- fit_stockseason(comp_dat = comp_ex, catch_dat = catch_ex, 
                      model_type = "integrated")
}
