% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2.R
\name{fit_stockseasonr}
\alias{fit_stockseasonr}
\title{Fit model}
\usage{
fit_stockseasonr(
  tmb_data,
  tmb_pars,
  tmb_map = NULL,
  tmb_random = NULL,
  fit_random = FALSE,
  ignore_fix = FALSE,
  model_specs,
  nlminb_loops = 1L,
  newton_loops = 0L
)
}
\arguments{
\item{tmb_data}{List of tmb input data generated in
\code{stockseasonr::make_inputs()}.}

\item{tmb_pars}{List of tmb input parameters generated in
\code{stockseasonr::make_inputs()}.}

\item{tmb_map}{Optional list of tmb input parameters to be mapped (i.e.
ignored) generated in \code{\link[=make_inputs]{make_inputs()}}.}

\item{tmb_random}{Optional list of tmb input parameters treated as random
effects (includes smooths and intercepts) generated in
\code{stockseasonr::make_inputs()}.}

\item{fit_random}{Logical (defaults to FALSE) specifying whether random
effects should be estimated.}

\item{ignore_fix}{Logical (defaults to FALSE) specifying whether a
fixed-effects only version of the model should be fit first to generate
initial parameters.}

\item{nlminb_loops}{How many times to run \code{\link[stats:nlminb]{stats::nlminb()}} optimization.
Sometimes restarting the optimizer at the previous best values aids
convergence. If the maximum gradient is still too large,
try increasing this to \code{2}.}

\item{newton_loops}{How many Newton optimization steps to try with
\code{\link[stats:optim]{stats::optimHess()}} after running \code{\link[stats:nlminb]{stats::nlminb()}}. Sometimes aids
convergence.}
}
\value{
Output from \code{\link[TMB:sdreport]{TMB::sdreport()}} and \link[summary(TMB:sdreport())]{summary(TMB::sdreport())}
}
\description{
Fit model
}
