% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-model-inputs_DEFUNCT.R
\name{gen_tmb}
\alias{gen_tmb}
\title{Prepare TMB inputs}
\usage{
gen_tmb(
  comp_dat,
  catch_dat = NULL,
  model_type = c("composition", "integrated"),
  random_walk = TRUE
)
}
\arguments{
\item{comp_dat}{Composition dataframe; each row represents a sampling event
(e.g. genetic samples collected from a given day and region).}

\item{catch_dat}{Catch data}

\item{model_type}{Character specifying whether inputs for a composition-only
or integrated model are being generated.}

\item{random_walk}{Logical (defaults to TRUE) specifying whether random
intercepts for year follow a random walk or are independent and identically
distributed.}
}
\value{
List of inputs (data and parameter lists) to pass to TMB and dataframes to
use in subsequent plotting.
}
\description{
Prepare TMB inputs
}
\examples{
dat_list <- gen_tmb(comp_dat = comp_ex, catch_dat = catch_ex, 
                    model_type = "integrated")
dat_list$data #data inputs for TMB
dat_list$pars #parameter inputs for TMB
}
