% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-model-inputs2.R
\name{make_inputs}
\alias{make_inputs}
\title{Prepare TMB inputs}
\usage{
make_inputs(
  abund_formula = NULL,
  comp_formula = NULL,
  comp_knots = NULL,
  abund_dat = NULL,
  comp_dat = NULL,
  abund_rint = NULL,
  comp_rint = NULL,
  pred_dat = NULL,
  model = c("negbin", "dirichlet", "integrated"),
  include_re_preds = FALSE
)
}
\arguments{
\item{abund_formula}{Model formula (excluding RIs) for abundance component.
Penalized splines are possible via mgcv with s(). See examples and
details below.}

\item{comp_formula}{Model formula (excluding RIs) for composition component.
Splines are possible via mgcv, but are not penalized (i.e. k must be
specified). See examples and details below.}

\item{comp_knots}{Optional list specifying knots location for composition
formula.}

\item{abund_dat}{Dataframe used to fit abundance component.}

\item{comp_dat}{Dataframe used to fit composition component.}

\item{abund_rint}{Name of column in abund_dat containing random intercepts.}

\item{comp_rint}{Name of column in comp_dat containing random intercepts.}

\item{pred_dat}{Dataframe used to generate predictions. Applies to both
abundance and composition components in integrated model.}

\item{model}{Character specifying whether abundance ("negbin"), composition
("dirichlet"), or both ("integrated") model components should be estimated.}

\item{include_re_preds}{Logical specifying whether predictions should include
random intercepts or not. In case of composition or integrated models this
requires an alternative model with a multivariate normal distribution that
can be unstable with large numbers of parameters.}
}
\value{
List of inputs (model_specs, tmb_data, tmb_pars, tmb_map and tmb_random) to
pass to TMB and dataframes to use in subsequent plotting.
}
\description{
Prepare TMB inputs
}
