% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ribbon_abund.R
\name{plot_ribbon_abund}
\alias{plot_ribbon_abund}
\title{Stock-specific abundance plot}
\usage{
plot_ribbon_abund(comp_dat, pred_dat_comp, mod_fit)
}
\arguments{
\item{comp_dat}{Composition dataframe; each row represents a sampling event
(e.g. genetic samples collected from a given day and region)}

\item{pred_dat_comp}{Dataframe of fixed effects composition predictions (i.e.
months and regions); generated using stockseasonr::gen_tmb()}

\item{mod_fit}{Parameter estimates generated by TMB::sdreport() and returned
by stockseasonr::fit_stockseason()}
}
\value{
Ribbon plot representing fixed effects predictions of seasonal trends in
stock-specific abundance, faceted by stock, with 95\% confidence intervals.
}
\description{
Stock-specific abundance plot
}
\examples{
in1 <- gen_tmb(comp_dat = comp_ex, catch_dat = catch_ex, 
               model_type = "integrated")
# model fitting will take several seconds
m1 <- fit_stockseason(comp_dat = comp_ex, catch_dat = catch_ex, 
                      model_type = "integrated")
plot_ribbon_abund(comp_dat = comp_ex, pred_dat_comp = in1$pred_dat_comp,
                  mod_fit = m1)

}
