% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_comp.R
\name{plot_stacked_comp}
\alias{plot_stacked_comp}
\title{Stacked composition plot}
\usage{
plot_stacked_comp(
  x_var,
  grouping_var,
  facet_var = NULL,
  ssdr,
  comp_dat,
  pred_dat
)
}
\arguments{
\item{x_var}{Explanatory variable within pred_dat to be plotted on x-axis
(must be continuous).}

\item{grouping_var}{Character or factor variable within pred_dat that
was used to generate composition estimates.}

\item{facet_var}{Optional character or factor variable within pred_dat
corresponding to categorical main effects in fitted model.}

\item{ssdr}{Parameter estimates generated by
stockseasonr::fit_stockseasonr()$ssdr.}

\item{comp_dat}{Composition dataframe; each row represents a sampling event
(e.g. genetic samples collected from a given day and region)}

\item{pred_dat}{Dataframe of fixed effects composition predictions}
}
\value{
Stacked ribbon plot representing trends in mean composition with optional
facets.
}
\description{
Stacked composition plot
}
\examples{
dum_pred <- expand.grid(
  month_n = seq(1, 12, by = 0.1),
  region = unique(comp_ex$region)
)
# model fitting will take several seconds
m1 <- fit_stockseasonr(
  comp_formula = agg ~ 1 + region + 
    s(month_n, bs = "tp", k = 4, m = 2) + 
    (1 | year),
  comp_dat = comp_ex,
  pred_dat = dum_pred,
  model = "dirichlet",
  random_walk = TRUE,
  fit = TRUE,
  nlminb_loops = 2, newton_loops = 1
)
plot_stacked_comp(x_var = "month_n", grouping_var = "agg", 
                  facet_var = "region", ssdr = m1$ssdr, comp_dat = comp_ex,
                  pred_dat = dum_pred)

}
