% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockseasonr-package.R
\docType{package}
\name{stockseasonr-package}
\alias{stockseasonr}
\alias{stockseasonr-package}
\title{stockseasonr: Seasonal Predictions of Stock Composition and Abundance}
\description{
An integrated model for estimating seasonal changes in abundance and stock composition. Designed and tested using fisheries data, estimates of total abundance assume a negative binomial distribution while composition estimates assume a Dirichlet-multinomial. stockseasonr is intended to generate predictions of composition and group-specific abundance. Group-specific abundance is calculated as the product of predicted total abundance and the probability of encountering a given group. It borrows significant functionality from glmmTMB and sdmTMB. Smooth terms can be incorporated using `s()` notation from mgcv and random intercepts following the familiar lme4 syntax. The full model is fit in TMB. Also includes basic plotting functions.
}
\author{
\strong{Maintainer}: Cameron Freshwater \email{cameron.freshwater@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Sean Anderson \email{sean.anderson@dfo-mpo.gc.ca}
}

}
\keyword{internal}
