% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_stockseason}
\alias{fit_stockseason}
\title{Fit model}
\usage{
fit_stockseason(catch_dat, comp_dat, silent = FALSE, nlminb_loops = 1)
}
\arguments{
\item{catch_dat}{Description here}

\item{comp_dat}{Description here}

\item{silent}{Logical}

\item{nlminb_loops}{How many times to run \code{\link[stats:nlminb]{stats::nlminb()}} optimization.
Sometimes restarting the optimizer at the previous best values aids
convergence. If the maximum gradient is still too large,
try increasing this to \code{2}.}
}
\value{
Output from \code{\link[TMB:sdreport]{TMB::sdreport()}}
}
\description{
Fit model
}
\examples{
# FIXME: make smaller/faster for example:
m <- fit_stockseason(catch_dat, comp_dat)
}
