% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_stockseason}
\alias{fit_stockseason}
\title{Fit model}
\usage{
fit_stockseason(
  comp_dat,
  catch_dat = NULL,
  model_type = c("composition", "integrated"),
  random_walk = TRUE,
  silent = FALSE,
  nlminb_loops = 1
)
}
\arguments{
\item{comp_dat}{Composition data}

\item{catch_dat}{Catch data}

\item{model_type}{Model type}

\item{random_walk}{Random walk or IID intercepts by year?}

\item{silent}{Logical}

\item{nlminb_loops}{How many times to run \code{\link[stats:nlminb]{stats::nlminb()}} optimization.
Sometimes restarting the optimizer at the previous best values aids
convergence. If the maximum gradient is still too large,
try increasing this to \code{2}.}
}
\value{
Output from \code{\link[TMB:sdreport]{TMB::sdreport()}}
}
\description{
Fit model
}
\examples{
m1 <- fit_stockseason(comp_ex, model_type = "composition")
m2 <- fit_stockseason(comp_ex, catch_ex, model_type = "integrated")
}
