% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2.R
\name{fit_stockseasonr}
\alias{fit_stockseasonr}
\title{Fit model}
\usage{
fit_stockseasonr(
  abund_formula = NULL,
  comp_formula = NULL,
  abund_dat = NULL,
  comp_dat = NULL,
  abund_offset = NULL,
  pred_dat = NULL,
  model = c("negbin", "dirichlet", "integrated"),
  random_walk = FALSE,
  fit = TRUE,
  silent = TRUE,
  nlminb_loops = 1L,
  newton_loops = 0L
)
}
\arguments{
\item{abund_formula}{Model formula for abundance component of model (fit with
negative binomial distribution (nbinom2)). Random intercepts (IID or RW)
are possible using lme4 syntax, e.g., + (1 | g) where g is a column of
class factor representing groups. Penalized splines are possible via mgcv
with s() (te not currently supported).}

\item{comp_formula}{Model formula for composition component of model (fit with
dirichlet distribution). Random intercepts (IID or RW) are possible using
lme4 syntax, e.g., + (1 | g) where g is a column of class factor
representing groups. Splines are possible via mgcv with s(), but unlike
abundance component these are unpenalized so knots must be specified with
k=.}

\item{abund_dat}{Dataframe containing abundance response (e.g. catch) and
covariates.}

\item{comp_dat}{Dataframe containing composition response in long format (i.e
column of class factor) and covariates.}

\item{abund_offset}{A numeric vector representing the model offset. Usually a
log transformed variable. \emph{Not included in any prediction.}}

\item{pred_dat}{Optional dataframe used to generate predictions. Note that
in integrated model this applies to both abundance and composition
components.}

\item{model}{String specifying whether abundance (negbin), composition
(dirichlet), or integrated model is fit.}

\item{random_walk}{Logical specifying whether random intercepts are IID or
random walk.}

\item{fit}{Logical specifying whether model should be fit in TMB. If FALSE
list of model inputs is returned.}

\item{silent}{Silent or include optimization details? Helpful to set to FALSE
for models that take a while to fit.}

\item{nlminb_loops}{How many times to run \code{\link[stats:nlminb]{stats::nlminb()}} optimization.
Sometimes restarting the optimizer at the previous best values aids
convergence. If the maximum gradient is still too large,
try increasing this to \code{2}.}

\item{newton_loops}{How many Newton optimization steps to try with
\code{\link[stats:optim]{stats::optimHess()}} after running \code{\link[stats:nlminb]{stats::nlminb()}}. Sometimes aids
convergence.}
}
\value{
List including model inputs as well as output from \code{\link[TMB:sdreport]{TMB::sdreport()}}
(if fit = TRUE).
}
\description{
Fit model
}
