% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ribbon_abund.R
\name{plot_ribbon_abund}
\alias{plot_ribbon_abund}
\title{Group-specific abundance plot}
\usage{
plot_ribbon_abund(
  x_var,
  grouping_var,
  colour_var = NULL,
  ssdr,
  comp_dat,
  pred_dat
)
}
\arguments{
\item{x_var}{Explanatory variable within pred_dat to be plotted on x-axis
(must be continuous).}

\item{grouping_var}{Character or factor variable within pred_dat that
was used to generate composition estimates.}

\item{colour_var}{Optional character or factor variable within pred_dat
corresponding to categorical main effects in fitted model.}

\item{ssdr}{Parameter estimates generated by
stockseasonr::fit_stockseasonr()$ssdr.}

\item{comp_dat}{Composition dataframe; each row represents a sampling event
(e.g. genetic samples collected from a given day and region)}

\item{pred_dat}{Dataframe of fixed effects composition predictions}
}
\value{
Ribbon plot representing fixed effects predictions of trends in group-
specific abundance, faceted by grouping variable, with 95\% confidence
intervals.
}
\description{
Group-specific abundance plot
}
\examples{
dum_pred <- expand.grid(
  month_n = seq(1, 12, by = 0.1),
  region = unique(comp_ex$region)
)
# model fitting will take several seconds
m2 <- fit_stockseasonr(
  abund_formula = catch ~ 1 +
    s(month_n, bs = "tp", k = 3, m = 2) +
    region +
    (1 | year),
  abund_dat = catch_ex,
  abund_offset = "offset",
  comp_formula = agg ~ 1 + region + 
    s(month_n, bs = "tp", k = 4, m = 2) + 
    (1 | year),
  comp_dat = comp_ex,
  pred_dat = dum_pred,
  model = "integrated",
  random_walk = TRUE,
  fit = TRUE,
  silent = FALSE,
  nlminb_loops = 2, newton_loops = 1
)
plot_ribbon_abund(x_var = "month_n", grouping_var = "agg", 
                  colour_var = "region", ssdr = m2$ssdr, comp_dat = comp_ex,
                  pred_dat = dum_pred)

}
