% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockseasonr-package.R
\docType{package}
\name{stockseasonr-package}
\alias{stockseasonr}
\alias{stockseasonr-package}
\title{stockseasonr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Cameron Freshwater \email{valid@email.com}

}
\keyword{internal}
